# Improve performance of `Owners` rule