… docs/markdown/Writing Plugins/common-plugin-tasks/plugin-upgrade-guide.md
… src/python/pants/core/goals/tailor_test.py